代理报表DTO修改

+++ src/main/java/com/dx/facade/report/resp/proxy/ProxyDayReportRespDTO.java
@@ -67,13 +67,13 @@ public class ProxyDayReportRespDTO implements Serializable {
     private BigDecimal depositAmount;
 
     @ApiModelProperty(value = "存款次数/会员存款次数")
-    private Integer depositTimes;
+    private Long depositTimes;
 
     @ApiModelProperty(value = "总取款/会员总取款")
     private BigDecimal withdrawAmount;
 
     @ApiModelProperty(value = "取款次数/会员取款次数")
-    private Integer withdrawTimes;
+    private Long withdrawTimes;
 
     @ApiModelProperty(value = "大额取款次数")
     private Integer bigWithdrawTimes;
